Refactored code with better structure and enums
