+++ packages/canary/src/styles.css
@@ -337,7 +337,7 @@
     --tag-background-red-02: var(--red/15);
   }
 
-  .light-std-standard {
+  .light-std-std {
     --background: 0 0% 100%;
     --foreground: 240 10% 3.9%;
     --card: 0 0% 100%;
@@ -385,7 +385,7 @@
     --diff-empty-content--: var(--background);
     --diff-hunk-content-color--: hsl(0, 0%, 100%);
   }
-  .dark-std-standard {
+  .dark-std-std {
     --background: var(--grey-6);
     --foreground: var(--grey-94);
     --card: var(--grey-10);
@@ -426,7 +426,7 @@
     --ai-button-stop-4: 188 97% 89%;
   }
 
-  .light-tri-standard {
+  .light-tri-std {
     --background: 0 0% 100%;
     --foreground: 20 14.3% 4.1%;
     --card: 0 0% 100%;
@@ -474,7 +474,7 @@
     --diff-empty-content--: var(--background);
     --diff-hunk-content-color--: hsl(0, 0%, 100%);
   }
-  .dark-tri-standard {
+  .dark-tri-std {
     --background: var(--grey-6);
     --foreground: var(--grey-94);
     --card: var(--grey-10);
@@ -515,7 +515,7 @@
     --ai-button-stop-4: 188 97% 89%;
   }
 
-  .light-pnd-standard {
+  .light-pnd-std {
     --background: 0 0% 100%;
     --foreground: 0 0% 3.9%;
     --card: 0 0% 100%;
@@ -563,7 +563,7 @@
     --diff-empty-content--: var(--background);
     --diff-hunk-content-color--: hsl(0, 0%, 100%);
   }
-  .dark-pnd-standard {
+  .dark-pnd-std {
     --background: var(--grey-6);
     --foreground: var(--grey-94);
     --card: var(--grey-10);
