+++ packages/views/src/components/theme-provider/types.ts
@@ -1,7 +1,23 @@
-export type ModeType = 'dark' | 'light' | 'system'
-export type ColorType = 'std' | 'tri' | 'pnd'
-export type ContrastType = 'standard' | 'low' | 'high'
+export enum ModeType {
+  Dark = 'dark',
+  Light = 'light',
+  System = 'system'
+}
+
+export enum ColorType {
+  Standard = 'std',
+  Tritanopia = 'tri',
+  ProtanopiaAndDeuteranopia = 'pnd'
+}
+
+export enum ContrastType {
+  Standard = 'std',
+  Low = 'low',
+  High = 'high'
+}
+
 export type FullTheme = `${ModeType}-${ColorType}-${ContrastType}`
+
 export type ThemeState = {
   theme: FullTheme
   setTheme: (theme: FullTheme) => void
