+++ packages/views/src/components/theme-selector/color-select.tsx
@@ -1,0 +1,50 @@
+import {
+  Label,
+  Select,
+  SelectContent,
+  SelectGroup,
+  SelectItem,
+  SelectLabel,
+  SelectSeparator,
+  SelectTrigger,
+  SelectValue
+} from '@harnessio/canary'
+
+import { ColorType, ContrastType, FullTheme, ModeType } from '../theme-provider/types'
+
+export function ColorSelect({
+  setTheme,
+  mode,
+  color,
+  contrast
+}: {
+  setTheme: (theme: FullTheme) => void
+  color: ColorType
+  mode: ModeType
+  contrast: ContrastType
+}) {
+  return (
+    <div>
+      <Label className="text-xs">Color</Label>
+      <Select
+        value={color}
+        onValueChange={(color: ColorType) => {
+          setTheme(`${mode}-${color}-${contrast}`)
+        }}
+      >
+        <SelectTrigger>
+          <SelectValue placeholder="Select a color theme" />
+        </SelectTrigger>
+        <SelectContent>
+          <SelectItem value={ColorType.Standard}>Standard</SelectItem>
+          <SelectSeparator />
+          <SelectGroup>
+            <SelectLabel>Vision Assistive</SelectLabel>
+            <SelectItem value={ColorType.Tritanopia}>Tritanopia</SelectItem>
+            <SelectItem value={ColorType.ProtanopiaAndDeuteranopia}>Protanopia & Deuteranopia</SelectItem>
+          </SelectGroup>
+        </SelectContent>
+      </Select>
+    </div>
+  )
+}
