+++ packages/views/src/components/theme-selector/mode-select.tsx
@@ -1,0 +1,36 @@
+import { Label, Select, SelectContent, SelectItem, SelectTrigger, SelectValue } from '@harnessio/canary'
+
+import { ColorType, ContrastType, FullTheme, ModeType } from '../theme-provider/types'
+
+export function ModeSelect({
+  setTheme,
+  mode,
+  color,
+  contrast
+}: {
+  setTheme: (theme: FullTheme) => void
+  mode: ModeType
+  color: ColorType
+  contrast: ContrastType
+}) {
+  return (
+    <div>
+      <Label className="text-xs">Mode</Label>
+      <Select
+        value={mode}
+        onValueChange={(mode: ModeType) => {
+          setTheme(`${mode}-${color}-${contrast}`)
+        }}
+      >
+        <SelectTrigger>
+          <SelectValue placeholder="Select a theme mode" />
+        </SelectTrigger>
+        <SelectContent>
+          <SelectItem value={ModeType.Light}>Light</SelectItem>
+          <SelectItem value={ModeType.Dark}>Dark</SelectItem>
+          <SelectItem value={ModeType.System}>System</SelectItem>
+        </SelectContent>
+      </Select>
+    </div>
+  )
+}
